Bis auf Repository soweit fertig
